requirements.txt
alembic==1.5.8
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
bcrypt==3.2.0
certifi==2020.6.20
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
fastapi-jwt-auth==0.5.0
fastapi-pagination==0.9.3
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
kombu==5.2.4
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.23.1
packaging==21.3
pandas==1.4.3
passlib==1.7.4
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2.binary==2.9.3
py==1.11.0
pycryptodome==3.15.0
pydantic==1.9.1
PyJWT==1.7.1
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
redis==4.3.3
requests==2.28.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
gunicorn
uvicorn==0.18.1
#uvloop==0.16.0
vine==5.0.0
watchfiles==0.15.0 
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
python-multipart==0.0.5
openpyxl===3.0.10
google-auth===2.11.0
apscheduler
httpx

schedule==1.2.0          # new 
boto3==1.28.60          # new 